A Wrapper for the BFS Packages download function

This wrapper allows to download functions from the BFS Package. Depending on the source either data or asset we use different functions. This function handles that distinction. It also only downloads if either teh data isn't at the given path or if there is a forced re-download

bfs_wrapper(id, dest_file, do_api_call = FALSE, type = "data")

Arguments

id

String. The BFS identifier (PX-ID for 'data' or Asset-ID for 'asset').

dest_file

String. The local file path where the data should be saved.

do_api_call

Bool. If TRUE, forces a new download even if the file exists locally. Defaults to FALSE.

type

String. The type of download: "data" for API table extraction or "asset" for direct file downloads (e.g., Excel/CSV).

Value

Bool Returns TRUE if the file exists or was successfully downloaded; FALSE otherwise.